requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
Pillow>=10.1.0
pyotp>=2.9.0
python-telegram-bot>=20.7
tweepy>=4.14.0
instagram-basic-display>=1.0.0
tiktok-api==0.10.5
redis>=5.0.1
gunicorn>=21.2.0
whitenoise>=6.6.0
requests>=2.31.0
cryptography>=41.0.7
django-storages==1.14.2
django-environ==0.11.2
django-rest-passwordreset==1.4.0
qrcode==7.4.2
social-auth-app-django==5.4.0
python-social-auth[django]
django-allauth==0.57.0
django-jazzmin==2.6.0
drf-yasg==1.21.7

# Subscription and Payment
stripe==7.11.0
celery==5.3.6
django-celery-beat==2.5.0 